+++ src/components/widget/tooltip/index.js
@@ -1,3 +1,0 @@
-import TooltipComponent from './component';
-
-export default TooltipComponent;